pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-ads>=28.4.1",
    "google-auth-oauthlib>=1.2.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "google-genai>=0.3.0",
    "google-adk>=0.0.1",
]
